Move admin problem handlers to handler package
